requirements.txt
asgiref==3.7.2
boto3==1.34.34
botocore==1.34.34
Django==5.0.1
django-storages==1.14.2
jmespath==1.0.1
python-dateutil==2.8.2
python-decouple==3.8
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
